Add files via upload
The 'Waste Kalyan' project is a hypothetical business case study that focuses on the reduction, reuse, and recycling of biodegradable waste into valuable resources such as manure, feedstock, and biofuel. The primary goal of this project is to protect the environment by implementing a comprehensive logistics system that prevents food vendors, restaurants, and households from contributing to the growing problem of waste accumulation in India's landfills. Through innovative solutions and sustainable practices, Waste Kalyan aims to create a more environmentally conscious and efficient waste management system for the benefit of all stakeholders involved.